requirements.txt
pds4-tools>=1.3
colour-demosaicing>=0.2
colour-science>=0.4
scikit-image>=0.21
Pillow>=10.0
numpy>=1.24
